+++ runtime/cere/src/lib.rs
@@ -437,14 +437,19 @@ impl pallet_babe::Config for Runtime {
 	type EpochDuration = EpochDuration;
 	type ExpectedBlockTime = ExpectedBlockTime;
 	type EpochChangeTrigger = pallet_babe::ExternalTrigger;
+
+	// In normal builds, respect session's disabled list.
+	#[cfg(not(feature = "try-runtime"))]
 	type DisabledValidators = Session;
+
+	// In try-runtime builds, bypass disabled validators to unblock simulation.
+	#[cfg(feature = "try-runtime")]
+	type DisabledValidators = ();
+
 	type MaxNominators = MaxNominatorRewardedPerValidator;
-
 	type KeyOwnerProof = sp_session::MembershipProof;
-
 	type EquivocationReportSystem =
 		pallet_babe::EquivocationReportSystem<Self, Offences, Historical, ReportLongevity>;
-
 	type WeightInfo = ();
 	type MaxAuthorities = MaxAuthorities;
 }
@@ -658,7 +663,7 @@ impl pallet_staking::Config for Runtime {
 	type MaxUnlockingChunks = ConstU32<32>;
 	type HistoryDepth = HistoryDepth;
 	type MaxControllersInDeprecationBatch = MaxControllersInDeprecationBatch;
-	type EventListeners = NominationPools;
+	type EventListeners = (NominationPools, DelegatedStaking);
 	type WeightInfo = pallet_staking::weights::SubstrateWeight<Runtime>;
 	type BenchmarkingConfig = StakingBenchmarkingConfig;
 	type NominationsQuota = pallet_staking::FixedNominationsQuota<{ MaxNominations::get() }>;
@@ -1459,6 +1464,14 @@ impl pallet_migrations::Config for Runtime {
 	type WeightInfo = pallet_migrations::weights::SubstrateWeight<Runtime>;
 }
 
+impl pallet_pool_withdrawal_fix::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type Currency = Balances;
+	type DelegationPalletConnector = DelegatedStaking;
+	type GovernanceOrigin = EnsureRoot<AccountId>;
+	type WeightInfo = ();
+}
+
 #[frame_support::runtime]
 mod runtime {
 	#[runtime::runtime]
@@ -1638,6 +1651,9 @@ mod runtime {
 	pub type MultiBlockMigrations = pallet_migrations;
 	#[runtime::pallet_index(52)]
 	pub type DelegatedStaking = pallet_delegated_staking;
+
+	#[runtime::pallet_index(53)]
+	pub type PoolWithdrawalFix = pallet_pool_withdrawal_fix::Pallet<Runtime>;
 }
 
 /// The address format for describing accounts.